Made InputOutputService interface
